+++ lib/src/model/user.dart
@@ -33,6 +33,16 @@ class User extends Equatable {
   /// How many comments the user contributed
   final int comments;
 
+  @JsonKey(name: 'following')
+
+  /// How many users this user follows
+  final int following;
+
+  @JsonKey(name: 'followers')
+
+  /// How many users follow this user
+  final int followers;
+
   @JsonKey(name: 'about')
 
   /// More about the user
@@ -45,6 +55,8 @@ class User extends Equatable {
   /// * [memberSince]: The user join date
   /// * [shaders]: The number of shaders
   /// * [comments]: The number of comments
+  /// * [following]: The number of users this user is following
+  /// * [followers]: The number of followers this user has
   /// * [about]: More about the user
   const User(
       {@required this.id,
@@ -52,18 +64,30 @@ class User extends Equatable {
       @required this.memberSince,
       this.shaders = 0,
       this.comments = 0,
+      this.following = 0,
+      this.followers = 0,
       this.about})
       : assert(id != null),
         assert(memberSince != null),
         assert(shaders != null),
-        assert(comments != null);
+        assert(comments != null),
+        assert(following != null),
+        assert(followers != null);
 
   @override
 
   /// The [List] of `props` (properties) which will be used to determine whether
   /// two [Equatables] are equal.
-  List<Object> get props =>
-      [id, picture, memberSince, shaders, comments, about];
+  List<Object> get props => [
+        id,
+        picture,
+        memberSince,
+        shaders,
+        comments,
+        following,
+        followers,
+        about
+      ];
 
   /// Creates a [User] from json map
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);