+++ lib/src/model/user.g.dart
@@ -15,6 +15,8 @@ User _$UserFromJson(Map<String, dynamic> json) {
         : DateTime.parse(json['memberSince'] as String),
     shaders: json['shaders'] as int,
     comments: json['comments'] as int,
+    following: json['following'] as int,
+    followers: json['followers'] as int,
     about: json['about'] as String,
   );
 }
@@ -25,5 +27,7 @@ Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
       'memberSince': instance.memberSince?.toIso8601String(),
       'shaders': instance.shaders,
       'comments': instance.comments,
+      'following': instance.following,
+      'followers': instance.followers,
       'about': instance.about,
     };